Add November newsletter
This also adds an image credit option for our blog headers,
and includes an Email Us link in the footer.

I've had a few publishers use the form to email us,
so I realized we should publish this information.
We probably want a contact page also,
but hopefully this will get less spam?