Add a license
All rights are reserved because this repository contains
our blog content, trademarks, and icons/images
that shouldn't be used without permission.

We are an open source company and this license isn't intended
to stop people from seeing how our landing pages and marketing
content is created or used.
If you want to use parts of this repository,
it's probably OK but please get in touch.