Add sponsorship landing page for SaaS
This is a first attempt at framing a SaaS/Sponsorship product.
It's targeted mainly at developers who might know what Open Collective is,
but we could make it more generic if we wanted.

This also has a few small refactors that I noticed while doing this,
and wanting to fix a couple typos.