Add Full publisher list
This is just copying the HTML from each of the existing pages,
because I couldn't find a nice way to pull that data dynamically.
If we have a good way do to this,
or store this data centrally in a way we can reuse that would be amazing.

Thinking something like:

```
data:
  publishers:
    frontend:
      - image:
        link:
      - image:
        link
...
```