Initial refactor of publisher landing pages
This moves the content onto the publisher page and out of the footer.
This is a reasonable refactor,
but feels odd without listing all our audiences.

I'd say we don't merge this one until we do more work to push it forward,
and build at least publisher pages for devops & security.